refactor: BillFilter folder
